user can load more tweets using EndlessRecyclerViewScrollListener

+++ app/src/main/java/com/codepath/apps/restclienttemplate/TwitterClient.java
@@ -53,7 +53,7 @@ public class TwitterClient extends OAuthBaseClient {
 		client.get(apiUrl, params, handler);
 	}
 
-	public void getNextPageOfTweets(JsonHttpResponseHandler handler, int maxId) {
+	public void getNextPageOfTweets(JsonHttpResponseHandler handler, long maxId) {
 		String apiUrl = getApiUrl("statuses/home_timeline.json");
 		// Can specify query string params directly or through RequestParams.
 		RequestParams params = new RequestParams();
